refactor: 版本比较全仓唯一实现（scripts/lib/versions.js），消除双实现语义漂移
updater.js 与 scripts/plugin-manager-update.js 各维护一份语义不同的
compareVersions（前者不去前导 v、只比前 3 段、段数不齐返回 NaN 导致漏报
更新；后者已完整实现且配套单测），companion-profile 还跨层 require 脚本
文件拿版本比较。收口为 scripts/lib/versions.js 一份实现：

- 数值分段比较（段数不限）、缺失段按 0、忽略前导 v、预发布语义与旧两版
  对全部真实调用形态（0.3.x 客户端版本、0.1.0-rc.N agent 版本、
  0.2.4-beta 插件版本）逐一比对结果一致，替换为零行为变更；
- plugin-manager-update.js 保留兼容性再导出（main.js / 既有单测沿用）；
- updater.js 顺带修复同批审计发现的错误处理缺陷：checkLatest 空输出
  无守卫抛 TypeError、overlayVersion/bundledVersion 用 require 读 JSON 的
  进程内脏读、并发 runNpm 时较早结束者清掉较晚者的 activeProc 引用、
  applyUpdate 成功后清理旧副本失败误报「更新失败」、saveSettings 失败
  无日志；
- check-syntax 入口清单补入新模块。
